pyproject.toml
[tool.poetry]
name = "crd2jsonschema"
version = "0.3.0"
description = "A utility to extract JSON Schema from a Kubernetes CRD"
authors = ["Alan J Castonguay <alan@tignis.com>"]
license = "Apache-2.0"
readme = "README.rst"
repository = "https://github.com/tignis/crd2jsonschema"
keywords = ["kubernetes", "crd", "jsonschema"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.1"

[tool.poetry.scripts]
crd2jsonschema = "crd2jsonschema:main"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
jsonschema = "^3.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
